lambda/pgvector-enabler/requirements.txt
psycopg2-binary
boto3
cfnresponse